refactor poms to their own package
